keep the original name of HnswGraphBuilder; add IHnswGraphBuilder for the interface (sorry)
