+++ lucene/core/src/java/org/apache/lucene/codecs/lucene95/Lucene95HnswVectorsWriter.java
@@ -45,7 +45,7 @@ import org.apache.lucene.util.hnsw.HnswGraph;
 import org.apache.lucene.util.hnsw.HnswGraph.NodesIterator;
 import org.apache.lucene.util.hnsw.NeighborArray;
 import org.apache.lucene.util.hnsw.OnHeapHnswGraph;
-import org.apache.lucene.util.hnsw.OnHeapHnswGraphBuilder;
+import org.apache.lucene.util.hnsw.HnswGraphBuilder;
 import org.apache.lucene.util.hnsw.OnHeapHnswGraphFactory;
 import org.apache.lucene.util.hnsw.RandomAccessVectorValues;
 import org.apache.lucene.util.packed.DirectMonotonicWriter;
@@ -439,7 +439,7 @@ public final class Lucene95HnswVectorsWriter extends KnnVectorsWriter {
                         docsWithField.cardinality(),
                         vectorDataInput,
                         byteSize);
-                OnHeapHnswGraphBuilder<byte[]> hnswGraphBuilder =
+                HnswGraphBuilder<byte[]> hnswGraphBuilder =
                     createHnswGraphBuilder(mergeState, fieldInfo, vectorValues, initializerIndex);
                 hnswGraphBuilder.setInfoStream(segmentWriteState.infoStream);
                 yield hnswGraphBuilder.build(vectorValues.copy());
@@ -451,7 +451,7 @@ public final class Lucene95HnswVectorsWriter extends KnnVectorsWriter {
                         docsWithField.cardinality(),
                         vectorDataInput,
                         byteSize);
-                OnHeapHnswGraphBuilder<float[]> hnswGraphBuilder =
+                HnswGraphBuilder<float[]> hnswGraphBuilder =
                     createHnswGraphBuilder(mergeState, fieldInfo, vectorValues, initializerIndex);
                 hnswGraphBuilder.setInfoStream(segmentWriteState.infoStream);
                 yield hnswGraphBuilder.build(vectorValues.copy());
@@ -483,7 +483,7 @@ public final class Lucene95HnswVectorsWriter extends KnnVectorsWriter {
     }
   }
 
-  private <T> OnHeapHnswGraphBuilder<T> createHnswGraphBuilder(
+  private <T> HnswGraphBuilder<T> createHnswGraphBuilder(
       MergeState mergeState,
       FieldInfo fieldInfo,
       RandomAccessVectorValues<T> floatVectorValues,
@@ -496,7 +496,7 @@ public final class Lucene95HnswVectorsWriter extends KnnVectorsWriter {
           fieldInfo.getVectorSimilarityFunction(),
           M,
           beamWidth,
-          OnHeapHnswGraphBuilder.randSeed);
+          HnswGraphBuilder.randSeed);
     }
 
     HnswGraph initializerGraph =
@@ -509,7 +509,7 @@ public final class Lucene95HnswVectorsWriter extends KnnVectorsWriter {
         fieldInfo.getVectorSimilarityFunction(),
         M,
         beamWidth,
-        OnHeapHnswGraphBuilder.randSeed,
+        HnswGraphBuilder.randSeed,
         initializerGraph,
         ordinalMapper);
   }
@@ -869,7 +869,7 @@ public final class Lucene95HnswVectorsWriter extends KnnVectorsWriter {
     private final int dim;
     private final DocsWithFieldSet docsWithField;
     private final List<T> vectors;
-    private final OnHeapHnswGraphBuilder<T> hnswGraphBuilder;
+    private final HnswGraphBuilder<T> hnswGraphBuilder;
 
     private int lastDocID = -1;
     private int node = 0;
@@ -906,7 +906,7 @@ public final class Lucene95HnswVectorsWriter extends KnnVectorsWriter {
               fieldInfo.getVectorSimilarityFunction(),
               M,
               beamWidth,
-              OnHeapHnswGraphBuilder.randSeed);
+              HnswGraphBuilder.randSeed);
       hnswGraphBuilder.setInfoStream(infoStream);
     }
