+++ lucene/core/src/java/org/apache/lucene/util/hnsw/HnswGraph.java
@@ -37,7 +37,7 @@ import org.apache.lucene.index.FloatVectorValues;
  * <h2>Hyperparameters</h2>
  *
  * <ul>
- *   <li><code>beamWidth</code> in {@link OnHeapHnswGraphBuilder} has the same meaning as <code>
+ *   <li><code>beamWidth</code> in {@link HnswGraphBuilder} has the same meaning as <code>
  *       efConst
  *       </code> in the paper. It is the number of nearest neighbor candidates to track while
  *       searching the graph for each newly inserted node.
