+++ lucene/core/src/java/org/apache/lucene/util/hnsw/IHnswGraphBuilder.java
@@ -18,19 +18,23 @@
 package org.apache.lucene.util.hnsw;
 
 import java.io.IOException;
-import org.apache.lucene.index.VectorEncoding;
-import org.apache.lucene.index.VectorSimilarityFunction;
 
-/** A factory for HnswGraphs. */
-public interface HnswGraphFactory {
-  public <T> IHnswGraphBuilder<T> createBuilder(
-      RandomAccessVectorValues<T> vectors,
-      VectorEncoding vectorEncoding,
-      VectorSimilarityFunction similarityFunction,
-      int M,
-      int beamWidth,
-      long seed)
-      throws IOException;
+/** A builder for HnswGraphs. */
+public interface IHnswGraphBuilder<T> {
+  /** Builds a graph from the given vectors */
+  public HnswGraph build(RandomAccessVectorValues<T> vectorsToAdd) throws IOException;
 
-  public HnswGraph create(int maxConnections);
+  /** Inserts a single doc with vector value to the graph. */
+  public void addGraphNode(int node, T value) throws IOException;
+
+  /** Inserts a single doc with vector value retrieved from the given supplier. */
+  default void addGraphNode(int node, RandomAccessVectorValues<T> values) throws IOException {
+    addGraphNode(node, values.vectorValue(node));
+  }
+
+  /**
+   * @return the graph built. It is okay to continue to add nodes after calling this; the reference
+   *     will continue to be valid.
+   */
+  public HnswGraph getGraph();
 }
