+++ lucene/core/src/test/org/apache/lucene/index/TestKnnGraph.java
@@ -18,7 +18,7 @@ package org.apache.lucene.index;
 
 import static com.carrotsearch.randomizedtesting.RandomizedTest.randomIntBetween;
 import static org.apache.lucene.search.DocIdSetIterator.NO_MORE_DOCS;
-import static org.apache.lucene.util.hnsw.OnHeapHnswGraphBuilder.randSeed;
+import static org.apache.lucene.util.hnsw.HnswGraphBuilder.randSeed;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -54,7 +54,7 @@ import org.apache.lucene.util.IOUtils;
 import org.apache.lucene.util.VectorUtil;
 import org.apache.lucene.util.hnsw.HnswGraph;
 import org.apache.lucene.util.hnsw.HnswGraph.NodesIterator;
-import org.apache.lucene.util.hnsw.OnHeapHnswGraphBuilder;
+import org.apache.lucene.util.hnsw.HnswGraphBuilder;
 import org.junit.After;
 import org.junit.Before;
 
@@ -63,7 +63,7 @@ public class TestKnnGraph extends LuceneTestCase {
 
   private static final String KNN_GRAPH_FIELD = "vector";
 
-  private static int M = OnHeapHnswGraphBuilder.DEFAULT_MAX_CONN;
+  private static int M = HnswGraphBuilder.DEFAULT_MAX_CONN;
 
   private Codec codec;
   private Codec float32Codec;
@@ -81,7 +81,7 @@ public class TestKnnGraph extends LuceneTestCase {
         new Lucene95Codec() {
           @Override
           public KnnVectorsFormat getKnnVectorsFormatForField(String field) {
-            return new Lucene95HnswVectorsFormat(M, OnHeapHnswGraphBuilder.DEFAULT_BEAM_WIDTH);
+            return new Lucene95HnswVectorsFormat(M, HnswGraphBuilder.DEFAULT_BEAM_WIDTH);
           }
         };
 
@@ -93,7 +93,7 @@ public class TestKnnGraph extends LuceneTestCase {
         new Lucene95Codec() {
           @Override
           public KnnVectorsFormat getKnnVectorsFormatForField(String field) {
-            return new Lucene95HnswVectorsFormat(M, OnHeapHnswGraphBuilder.DEFAULT_BEAM_WIDTH);
+            return new Lucene95HnswVectorsFormat(M, HnswGraphBuilder.DEFAULT_BEAM_WIDTH);
           }
         };
 
@@ -104,7 +104,7 @@ public class TestKnnGraph extends LuceneTestCase {
           new Lucene95Codec() {
             @Override
             public KnnVectorsFormat getKnnVectorsFormatForField(String field) {
-              return new Lucene95HnswVectorsFormat(M, OnHeapHnswGraphBuilder.DEFAULT_BEAM_WIDTH);
+              return new Lucene95HnswVectorsFormat(M, HnswGraphBuilder.DEFAULT_BEAM_WIDTH);
             }
           };
     }
@@ -116,7 +116,7 @@ public class TestKnnGraph extends LuceneTestCase {
 
   @After
   public void cleanup() {
-    M = OnHeapHnswGraphBuilder.DEFAULT_MAX_CONN;
+    M = HnswGraphBuilder.DEFAULT_MAX_CONN;
   }
 
   /** Basic test of creating documents in a graph */
