+++ lucene/core/src/test/org/apache/lucene/util/hnsw/HnswGraphTestCase.java
@@ -291,7 +291,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     int nDoc = 100;
     similarityFunction = VectorSimilarityFunction.DOT_PRODUCT;
     RandomAccessVectorValues<T> vectors = circularVectorValues(nDoc);
-    HnswGraphBuilder<T> builder =
+    IHnswGraphBuilder<T> builder =
         factory.createBuilder(
             vectors, getVectorEncoding(), similarityFunction, 10, 100, random().nextInt());
     HnswGraph hnsw = builder.build(vectors.copy());
@@ -342,7 +342,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     int nDoc = 100;
     RandomAccessVectorValues<T> vectors = circularVectorValues(nDoc);
     similarityFunction = VectorSimilarityFunction.DOT_PRODUCT;
-    HnswGraphBuilder<T> builder =
+    IHnswGraphBuilder<T> builder =
         factory.createBuilder(
             vectors, getVectorEncoding(), similarityFunction, 16, 100, random().nextInt());
     HnswGraph hnsw = builder.build(vectors.copy());
@@ -386,7 +386,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     int nDoc = 100;
     RandomAccessVectorValues<T> vectors = circularVectorValues(nDoc);
     similarityFunction = VectorSimilarityFunction.DOT_PRODUCT;
-    HnswGraphBuilder<T> builder =
+    IHnswGraphBuilder<T> builder =
         factory.createBuilder(
             vectors, getVectorEncoding(), similarityFunction, 16, 100, random().nextInt());
     HnswGraph hnsw = builder.build(vectors.copy());
@@ -532,7 +532,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     int nDoc = 500;
     similarityFunction = VectorSimilarityFunction.DOT_PRODUCT;
     RandomAccessVectorValues<T> vectors = circularVectorValues(nDoc);
-    HnswGraphBuilder<T> builder =
+    IHnswGraphBuilder<T> builder =
         factory.createBuilder(
             vectors, getVectorEncoding(), similarityFunction, 16, 100, random().nextInt());
     HnswGraph hnsw = builder.build(vectors.copy());
@@ -602,7 +602,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
         RandomizedTest.randomFrom(VectorSimilarityFunction.values());
     RandomAccessVectorValues<T> vectors = vectorValues(size, dim);
 
-    HnswGraphBuilder<T> builder =
+    IHnswGraphBuilder<T> builder =
         factory.createBuilder(
             vectors, getVectorEncoding(), similarityFunction, M, M * 2, random().nextLong());
     HnswGraph hnsw = builder.build(vectors.copy());
@@ -627,7 +627,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     };
     AbstractMockVectorValues<T> vectors = vectorValues(values);
     // First add nodes until everybody gets a full neighbor list
-    HnswGraphBuilder<T> builder =
+    IHnswGraphBuilder<T> builder =
         factory.createBuilder(
             vectors, getVectorEncoding(), similarityFunction, 2, 10, random().nextInt());
     // node 0 is added by the builder constructor
@@ -683,7 +683,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     };
     AbstractMockVectorValues<T> vectors = vectorValues(values);
     // First add nodes until everybody gets a full neighbor list
-    HnswGraphBuilder<T> builder =
+    IHnswGraphBuilder<T> builder =
         factory.createBuilder(
             vectors, getVectorEncoding(), similarityFunction, 1, 10, random().nextInt());
     RandomAccessVectorValues<T> vectorsCopy = vectors.copy();
@@ -715,7 +715,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     };
     AbstractMockVectorValues<T> vectors = vectorValues(values);
     // First add nodes until everybody gets a full neighbor list
-    HnswGraphBuilder<T> builder =
+    IHnswGraphBuilder<T> builder =
         factory.createBuilder(
             vectors, getVectorEncoding(), similarityFunction, 1, 10, random().nextInt());
     RandomAccessVectorValues<T> vectorsCopy = vectors.copy();
@@ -749,7 +749,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     int dim = atLeast(10);
     AbstractMockVectorValues<T> vectors = vectorValues(size, dim);
     int topK = 5;
-    HnswGraphBuilder<T> builder =
+    IHnswGraphBuilder<T> builder =
         factory.createBuilder(
             vectors, getVectorEncoding(), similarityFunction, 10, 30, random().nextLong());
     HnswGraph hnsw = builder.build(vectors.copy());
